Some explicit casting to remove warnings

+++ src/OpenMesh/Tools/Decimater/ModAspectRatioT.cc
@@ -109,7 +109,8 @@ void ModAspectRatioT<MeshT>::initialize() {
     typename Mesh::Point& p1 = mesh_.point(*(++fv_it));
     typename Mesh::Point& p2 = mesh_.point(*(++fv_it));
 
-    mesh_.property(aspect_, *f_it) = static_cast<typename Mesh::Scalar>(1.0) / aspectRatio(p0, p1, p2);
+    mesh_.property(aspect_, *f_it) = static_cast<float>(
+        static_cast<typename Mesh::Scalar>(1.0) / aspectRatio(p0, p1, p2));
   }
 }
 
@@ -128,7 +129,7 @@ void ModAspectRatioT<MeshT>::preprocess_collapse(const CollapseInfo& _ci) {
       typename Mesh::Point& p1 = mesh_.point(*(++fv_it));
       typename Mesh::Point& p2 = mesh_.point(*(++fv_it));
 
-      mesh_.property(aspect_, fh) = static_cast<typename Mesh::Scalar>(1.0) / aspectRatio(p0, p1, p2);
+      mesh_.property(aspect_, fh) = static_cast<float>(static_cast<typename Mesh::Scalar>(1.0) / aspectRatio(p0, p1, p2));
     }
   }
 }
